% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload_package}
\alias{reload_package}
\title{Perform all steps to reload the package}
\usage{
reload_package()
}
\description{
Perform all steps to reload the package
}
\details{
The function performs the following steps:
\enumerate{
\item Compiles the stan code, if changed
\item Updates the documentation
\item Installs the package
See this \href{https://gist.github.com/jamesmbaazam/6587ac77f1e37ed7373bced0cba3931d}{GitHub gist} for more details. #nolint
}
}
\examples{
\dontrun{
  reload_package()
}
}
