% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{StanDevUtils-package}
\alias{StanDevUtils}
\alias{StanDevUtils-package}
\title{StanDevUtils: Helper Functions to Facilitate R Package Development with Stan}
\description{
Provides wrappers around rinse and repeat functions to streamline R package development with stan.
}
\author{
\strong{Maintainer}: James M. Azam \email{james.azam@lshtm.ac.uk} [copyright holder]

}
\keyword{internal}
